выполнено задание модуляция
